Comment more cases of bad sampleCount truncation, and even worse...
  Fix progress indicator again: must do a floating-point division...
  A little more type agnosticism
  Remove more unnecessary casts to sampleCount
  Fix more narrowings of sampleCount

+++ src/TrackPanel.cpp
@@ -4383,7 +4383,7 @@ void TrackPanel::HandleSampleEditingClick( wxMouseEvent & event )
 
       //Get a sample  from the track to do some tricks on.
       mDrawingTrack->Get((samplePtr)sampleRegion, floatSample,
-                                       (int)mDrawingStartSample - SMOOTHING_KERNEL_RADIUS - SMOOTHING_BRUSH_RADIUS,
+                                       mDrawingStartSample - SMOOTHING_KERNEL_RADIUS - SMOOTHING_BRUSH_RADIUS,
                                        sampleRegionSize);
       int i, j;
